+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- A new parameter set for the Sobol'-G function from Sun et al. (2022).
 - The 6-dimensional undamped non-linear oscillator function for reliability
   analysis exercises.
 - The M-dimension Sobol'-Levitan function from Sobol' and Levitan (1999) for